requirements.txt
numpy==1.26.4
matplotlib==3.8.3
pyyaml==6.0.1
streamlit==1.33.0
pandas==2.0.3